+++ docs/src/changelog.md
@@ -2,6 +2,10 @@
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+- ![](https://img.shields.io/badge/new%20feature-green.svg) added semi-elliptic membership functions
+
 ## v0.1.2 -- 2023-03-12
 
 [view release on GitHub](https://github.com/lucaferranti/FuzzyLogic.jl/releases/tag/v0.1.2)